refactor(#103): move config.ts
